fix: configure Multicall3 address on viem PublicClient
viem's multicall() requires a chain with contracts.multicall3 configured.
Extracted shared getClient() into rpc.ts with defineChain that sets the
deterministic Multicall3 address (0xcA11bde...CA11).
